Made default Stack Overflow hook fail if configASSERT is defined
I beleive that the default implementation should fail if the user defines configASSERT. I believe this is a reasonable addition, given the fact configCHECK_FOR_STACK_OVERFLOW is already enabled by default. I hope it could save someone a lot of headache - as it happened to me when I overlooked my stack overflowing.