FW3: Comment out cmake path template definitions
CMSIS-FreeRTOS's CMakeLists.txt requires some variables to be defined by the user.

The file includes uncommented example commands such as:
  set(FREERTOS_PORT "" CACHE STRING "FreeRTOS port name")
with the expectation that the user will set these variables either in this file
or by a parent.

FW3 will define these by the parent, so these statements need to be
removed so they will not replace the desired values.